stream: add simple stream implementation
This initial Stream implementation adds a very simple wrapper around
an Ably channel. Messages received on the channel are surfaced to the
caller via a subscription callback.

The Stream is an EventEmitter and exposes its state via events.
Additionally the stream can be pauses, resumed and disposed. This
will allow the Models SDK to manipulate the stream/channel according to
the model's lifecycle.

While not yet particularly useful, the idea is that the Stream will
handle:
- message de-duplication
- message re-ordering
- detect discontinuities (due to connection state recovery as well as
  due to failed connector publishes)