stream: constructor args
The Stream constructor now requires that a logger instance is provided
rather than logger configuration (level). This allows the same logger
instance to be shared through the application.

The ably instance must also be provided on the stream options object.

The filter option is temporarily removed to aid subsequent changes which
use the channel to identify the stream.