demo: render optimistic comments with transparency
This commit attaches an `optimistic` field to comments which is set according to the `confirmed` property on the event passed to the update function. Optimistic comments are rendered with reduced opacity to indicate to the user that they are not yet confirmed.

This commit also improves the way the user is passed to the client by loading it in the top-level server-rendered component.