docs: refactor types to better expose API surface
This change moves types previously colocated with the module
implementation to a separate location, which is the re-exported from
the library entrypoint. This makes it much cleaner and easier to expose
exactly the API surface that we want to expose to users. Additionally
this makes the autogenerated documentation much easier to navigate.