mutations: apply optimistic events before mutation
This commit updates the mutation handler to apply optimistic events via the onEvents callback before the registered mutation method is invoked. This allows us to preserve the behaviour by which the caller can await the mutation method result when invoking the mutation handler while also allowing us to apply optimistic updates without having to wait for the (potentially slow) mutation method to complete.

This approach requires updating the error handling logic to catch both mutation method errors and errors from the onEvents callback. If errors occur in both, they are combined using an AggregateError. This necessitates updating the compile target to ES2021, so if we decide to support a lower version we will need to add a polyfill.