mutations: apply optimistic update before mutate
This change refactors the mutations registry to apply optimistic events
before the mutation method is invoked. This requires an additional
change to ensure that the optimistic events are rolled back if the
mutation method fails. Additionally, we no longer return a promise for
the optimistic update as it is always applied before the mutation
handler resolves.

- fix mutations registry implementation
- revert events on confirmation timeout
- fix handle uncaught errors and rejections
- no longer return optimistic promise
- refactor to class methods
- tidy up