mutations: set custom comparator on mutation opts
This commit removes the ability to pass in a custom comparator to the mutation function `$expect` method. Instead, it should be configured on the MutationOptions set on the mutation in `$register.$mutate`.

Setting custom comparator on a per-invocation basis will be supported in a subsequent commit with a new API design.