mutations: refactor mutations registry
This change updates the mutations API to support providing a set of
parameters to a mutation function when it is invoked via $expect.
Instead of passing the expected events directly, they are included on
the `events` key on the parameters. This supports also passing
invocation-specific mutation options, to e.g. configure a timeout on a
per-invocation basis.

This PR also improves the quality of the MutationsRegistry
implementation and makes the typing must easier to grok.