Add SlidingWindow to reorder incoming events
Add a sliding window to the Stream that will buffer the events for
config amount of time. It uses this buffer window to reorder events and
pass them to the subscription them once the timeout expires.

All events before the expiring event will be passed to the subscription.

For example:
- Receiving [e2, e1]
- Will be reordered to [e1, e2]

Assuming that e2 expires from the window before e1 (because it was added
to the window before e1) then both e1 and e2 will be passed to the
subscription.