Events represent state change, not total state
Optimistic and Confirmed events represent the state changes, and not the
total state after the change. Add, edit, and delete comment events only
contain the comment that was added, edited, or deleted.

Rely on client-side ordering based on createdAt time to sort the
comments correctly.

Add the date strings to the comments component.

Also:
- fix a bug where model.subscribe(...) could be called on a disposed
  model when using hot-reload.