Update error handling for errors from user functions
Update the error handling for errors that are thrown from user provided
functions.

There are three functions where we call into the user's code:
1. subscribe callback
2. sync function
3. merge function

If the sync function fails, we bubble the error back up to the caller,
which in most cases is the user, but in some resync cases is the
ModelsSDK.

For the merge function, if the user throws an error from merge, we try
and resync the model. If this resync operation fails, then we set the
model to errored, and expose an error to the `model.on(...)` method.

If the subscribe callback throws an error we ignore it.