Use persistLast to check for expired messages
Assume that persist last is enabled on the channel, and use the
sequenceID of the last message on the channel to calculate if the
channel has had no new changes on it, or if the channel messages have
expired.

If the messages have expired then throw an error. If there are no new
messages on the channel then continue processing knowing the state is up
to date.

Note: this change means that models requires channels with persist last
enabled on them, else the model sync will fail as it cannot compute the
most recent message on the channel.