model: use single sync func type parameter
This change updates the model to accept a single type parameter
determined from the registered sync function, instead of separately
definining the sync function parameters and model state type, both of
which can be inferred from the sync function definition.

While there is some complexity in the type definitions and ugliness
with e.g. having to refer to the state type as ExtractData<S> instead of
as a single generic type parameter T, the benefit is that the user gets
much stronger type inference when using the library. The user does not
need to explicitly pass the type params to the model; instead, it will
be correctly inferred from the sync function registration. Additionally,
any mismatch between the type of the data field returned by the sync
function and the type returned by the merge function will also be
caught. This is reflected in the updated tests. Full type inference is
maintained on the model data itself either via
models.data.confirmed/optimistic or via the model.subscribe() callback.