Resume on reconnect
The resume flow was never being triggered because the resync had an
infinite retry. Once onStreamMessage(err) was called, this triggers are
retryable resync loop. Only once that loop has exited do we pause the
model, and only once the model is paused do we try and resume using the
channel history.

We want to keep the behaviour of calling resync when onStreamMessage
receives an error, incase the error is caused by processing that
message. In which case a full resync is useful.

The time taken to trigger the pause/resume flow is 2mins + retryStrategy
total time.

This commit adds a limit to the number of retries that resync will do
both by default, and in the example app.